Make Operation enum pub(crate)
Rename Remove to RemoveValue and Empty to RemoveEntry to match
WriteHandle methods.
Deletes pending() as it leaked the Operation enum.
