+++ src/write.rs
@@ -336,38 +336,12 @@ where
         self
     }
 
-    /// Gives the sequence of operations that have not yet been applied.
-    ///
-    /// Note that until the *first* call to `refresh`, the sequence of operations is always empty.
-    ///
-    /// ```
-    /// # use evmap::Operation;
-    /// let x = ('x', 42);
-    ///
-    /// let (r, mut w) = evmap::new();
-    ///
-    /// // before the first refresh, no oplog is kept
-    /// w.refresh();
-    ///
-    /// assert_eq!(w.pending(), &[]);
-    /// w.insert(x.0, x);
-    /// assert_eq!(w.pending(), &[Operation::Add(x.0, x)]);
-    /// w.refresh();
-    /// w.remove(x.0, x);
-    /// assert_eq!(w.pending(), &[Operation::Remove(x.0, x)]);
-    /// w.refresh();
-    /// assert_eq!(w.pending(), &[]);
-    /// ```
-    pub fn pending(&self) -> &[Operation<K, V>] {
-        &self.oplog[self.swap_index..]
-    }
-
     /// Refresh as necessary to ensure that all operations are visible to readers.
     ///
     /// `WriteHandle::refresh` will *always* wait for old readers to depart and swap the maps.
     /// This method will only do so if there are pending operations.
     pub fn flush(&mut self) -> &mut Self {
-        if !self.pending().is_empty() {
+        if !self.oplog[self.swap_index..].is_empty() {
             self.refresh();
         }
 
@@ -440,7 +414,7 @@ where
     ///
     /// The updated value-bag will only be visible to readers after the next call to `refresh()`.
     pub fn remove_value(&mut self, k: K, v: V) -> &mut Self {
-        self.add_op(Operation::Remove(k, v))
+        self.add_op(Operation::RemoveValue(k, v))
     }
 
     /// Remove the value-bag for the given key.
@@ -455,7 +429,7 @@ where
     ///
     /// The value-bag will only disappear from readers after the next call to `refresh()`.
     pub fn remove_entry(&mut self, k: K) -> &mut Self {
-        self.add_op(Operation::Empty(k))
+        self.add_op(Operation::RemoveEntry(k))
     }
 
     /// Purge all value-bags from the map.
@@ -593,7 +567,7 @@ where
                     .or_insert_with(Values::new)
                     .push(unsafe { value.shallow_copy() }, hasher);
             }
-            Operation::Empty(ref key) => {
+            Operation::RemoveEntry(ref key) => {
                 #[cfg(not(feature = "indexed"))]
                 inner.data.remove(key);
                 #[cfg(feature = "indexed")]
@@ -608,7 +582,7 @@ where
                     inner.data.swap_remove_index(index);
                 }
             }
-            Operation::Remove(ref key, ref value) => {
+            Operation::RemoveValue(ref key, ref value) => {
                 if let Some(e) = inner.data.get_mut(key) {
                     // remove a matching value from the value set
                     // safety: this is fine
@@ -671,7 +645,7 @@ where
                     .or_insert_with(Values::new)
                     .push(value, hasher);
             }
-            Operation::Empty(key) => {
+            Operation::RemoveEntry(key) => {
                 #[cfg(not(feature = "indexed"))]
                 inner.data.remove(&key);
                 #[cfg(feature = "indexed")]
@@ -686,7 +660,7 @@ where
                     inner.data.swap_remove_index(index);
                 }
             }
-            Operation::Remove(key, value) => {
+            Operation::RemoveValue(key, value) => {
                 if let Some(e) = inner.data.get_mut(&key) {
                     // find the first entry that matches all fields
                     e.swap_remove(&value);
@@ -752,5 +726,26 @@ where
     type Target = ReadHandle<K, V, M, S>;
     fn deref(&self) -> &Self::Target {
         &self.r_handle
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::new;
+
+    #[test]
+    fn flush_noblock() {
+        let x = ('x', 42);
+
+        let (r, mut w) = new();
+        w.insert(x.0, x);
+        w.refresh();
+        assert_eq!(r.get(&x.0).map(|rs| rs.len()), Some(1));
+
+        // pin the epoch
+        let _map = r.read();
+        // refresh would hang here, but flush won't
+        assert!(w.oplog[w.swap_index..].is_empty());
+        w.flush();
     }
 }
